Started on scanning for types
