+++ src/code/llvm/mod.rs
@@ -1,25 +1,28 @@
-use std::{convert::TryFrom, ops::Deref, path::Path, process::Command};
+pub mod compile;
+pub mod types;
+use std::{convert::TryFrom, ops::Deref};
 
 use crate::{
     ast::{Ast, FunProto},
     lex::Op,
     types::Container,
-    CompileOpts, OutFormat, Type,
+    Type,
 };
 use hashbrown::HashMap;
 use inkwell::{
     builder::Builder,
     context::Context,
     module::Module,
-    passes::PassManager,
-    targets::{CodeModel, FileType, InitializationConfig, RelocMode, Target, TargetMachine},
     types::{AnyType, AnyTypeEnum, BasicType, BasicTypeEnum, StructType},
     values::{AnyValue, AnyValueEnum, BasicValue, BasicValueEnum, FunctionValue, PointerValue},
-    IntPredicate, OptimizationLevel,
+    IntPredicate,
 };
 
 /// The `Compiler` struct is used to generate an executable with LLVM from the parsed AST.
 pub struct Compiler<'c> {
+    /// The name of the currently compiled module
+    name: String,
+
     /// The LLVM context
     ctx: &'c Context,
 
@@ -53,8 +56,9 @@ pub struct Compiler<'c> {
 
 impl<'c> Compiler<'c> {
     /// Create a new `Compiler` from an LLVM context struct
-    pub fn new(ctx: &'c Context) -> Self {
+    pub fn new(ctx: &'c Context, name: String) -> Self {
         Self {
+            name,
             ctx,
             build: ctx.create_builder(),
             module: ctx.create_module("spark_llvm_module"),
@@ -137,31 +141,6 @@ impl<'c> Compiler<'c> {
             Ok(fun)
         }
     }
-
-    /// Get a struct type from the given path
-    pub fn get_struct(&self, name: impl AsRef<str>) -> Option<(StructType<'c>, Container)> {
-        let name = name.as_ref();
-        self.struct_types.get(name).cloned()
-    }
-
-    /// Get a union type from the given path
-    pub fn get_union(&self, name: impl AsRef<str>) -> Option<(StructType<'c>, Container)> {
-        let name = name.as_ref();
-        self.union_types.get(name).cloned()
-    }
-
-    /// Get a typedef'd type from the given path
-    pub fn get_typedef(&self, name: impl AsRef<str>) -> Option<Type> {
-        let name = name.as_ref();
-        self.typedefs.get(name).cloned()
-    }
-
-    /// Get a struct type from the given path
-    pub fn get_fun(&self, name: impl AsRef<str>) -> Option<(FunctionValue<'c>, FunProto)> {
-        let name = name.as_ref();
-        self.funs.get(name).cloned()
-    }
-
     /// Build an alloca for a variable in the current function
     fn entry_alloca(&self, name: &str, ty: BasicTypeEnum<'c>) -> PointerValue<'c> {
         let entry_builder = self.ctx.create_builder();
@@ -798,140 +777,5 @@ impl<'c> Compiler<'c> {
             })
             .collect::<Vec<_>>()
     }
-
-    /// Generate all code for a LLVM module and return it
-    pub fn finish(mut self, ast: Vec<Ast>) -> Module<'c> {
-        let ast = self.get_decls(ast);
-        //let ast = self.get_fn_protos(ast);
-        for node in ast {
-            self.gen(&node, false);
-        }
-        self.module
-    }
-
-    /// Compile the code into an executable file
-    pub fn compile(self, ast: Vec<Ast>, opts: CompileOpts) {
-        const LINKER: &str = "C:\\Program Files (x86)\\Microsoft Visual Studio\\2019\\BuildTools\\VC\\Tools\\MSVC\\14.28.29910\\bin\\Hostx64\\x64\\link.exe";
-        use std::process::Stdio;
-
-        let module = self.finish(ast);
-
-        module
-            .verify()
-            .unwrap_or_else(|e| panic!("Failed to verify the LLVM module: {}", e));
-
-        let fpm: PassManager<Module<'c>> = PassManager::create(());
-
-        match opts.opt_lvl {
-            crate::OptLvl::Debug => (),
-            crate::OptLvl::Medium => {
-                fpm.add_demote_memory_to_register_pass();
-                fpm.add_promote_memory_to_register_pass();
-                fpm.add_constant_merge_pass();
-                fpm.add_instruction_combining_pass();
-                fpm.add_global_optimizer_pass();
-            }
-            crate::OptLvl::Aggressive => {
-                fpm.add_demote_memory_to_register_pass();
-                fpm.add_promote_memory_to_register_pass();
-                fpm.add_constant_merge_pass();
-                fpm.add_instruction_combining_pass();
-                fpm.add_global_optimizer_pass();
-
-                fpm.add_loop_rotate_pass();
-                fpm.add_argument_promotion_pass();
-                fpm.add_function_inlining_pass();
-                fpm.add_memcpy_optimize_pass();
-                fpm.add_loop_deletion_pass();
-                fpm.add_loop_vectorize_pass();
-                fpm.add_constant_propagation_pass();
-                fpm.add_simplify_lib_calls_pass();
-                fpm.add_strip_symbol_pass();
-            }
-        }
-
-        match opts.output_ty {
-            OutFormat::IR => module.print_to_file(opts.out_file).unwrap(),
-            other => {
-                Target::initialize_all(&InitializationConfig::default());
-                let opt = OptimizationLevel::Aggressive;
-                let reloc = RelocMode::Default;
-                let model = CodeModel::Default;
-                let target = Target::from_triple(&TargetMachine::get_default_triple()).unwrap();
-                let machine = target
-                    .create_target_machine(
-                        &TargetMachine::get_default_triple(),
-                        &TargetMachine::get_host_cpu_name().to_str().unwrap(),
-                        &TargetMachine::get_host_cpu_features().to_str().unwrap(),
-                        opt,
-                        reloc,
-                        model,
-                    )
-                    .unwrap();
-
-                machine.add_analysis_passes(&fpm);
-
-                match other {
-                    OutFormat::Asm => machine
-                        .write_to_file(&module, FileType::Assembly, &opts.out_file)
-                        .unwrap(),
-                    OutFormat::Obj => machine
-                        .write_to_file(&module, FileType::Object, &opts.out_file)
-                        .unwrap(),
-                    OutFormat::Lib => {
-                        let obj = opts.out_file.with_extension("obj");
-                        machine
-                            .write_to_file(&module, FileType::Object, &obj)
-                            .unwrap();
-
-                        let out = format!("/OUT:{}", opts.out_file.display());
-                        let mut args = vec!["/LIB", obj.to_str().unwrap(), "/NOLOGO", out.as_str()];
-                        args.extend(opts.libraries.iter().map(|s| s.as_str())); //Add all linked libraries
-
-                        let cmd = Command::new(Path::new(LINKER))
-                            .args(args)
-                            .stderr(Stdio::piped())
-                            .stdout(Stdio::piped())
-                            .spawn()
-                            .unwrap(); //Link the file into a library
-                        println!(
-                            "{}",
-                            String::from_utf8(cmd.wait_with_output().unwrap().stdout).unwrap()
-                        );
-
-                        std::fs::remove_file(obj).unwrap();
-                    }
-                    OutFormat::Exe => {
-                        let obj = opts.out_file.with_extension("obj");
-                        machine
-                            .write_to_file(&module, FileType::Object, &obj)
-                            .unwrap();
-
-                        let out = format!("/OUT:{}", opts.out_file.display());
-                        let mut args = vec![
-                            obj.to_str().unwrap(),
-                            "/ENTRY:main",
-                            "/SUBSYSTEM:console",
-                            "/NOLOGO",
-                            out.as_str(),
-                        ];
-                        args.extend(opts.libraries.iter().map(|s| s.as_str())); //Add all linked libraries
-
-                        let cmd = Command::new(Path::new(LINKER))
-                            .args(args)
-                            .stderr(Stdio::piped())
-                            .stdout(Stdio::piped())
-                            .spawn()
-                            .unwrap(); //Link the file into a library
-                        println!(
-                            "{}",
-                            String::from_utf8(cmd.wait_with_output().unwrap().stdout).unwrap()
-                        );
-                        std::fs::remove_file(obj).unwrap();
-                    }
-                    _ => unreachable!(),
-                }
-            }
-        }
-    }
+    
 }
