+++ src/main.rs
@@ -6,7 +6,6 @@ pub mod types;
 use std::{
     panic::PanicInfo,
     path::PathBuf,
-    process::{Command, Stdio},
     str::FromStr,
 };
